feat: remove v0 features
