+++ DiscordBotsList.Api.Tests/UnitTests.cs
@@ -21,55 +21,19 @@ namespace DiscordBotsList.Api.Tests
 
     public class UnitTests
     {
-        private readonly AuthV1DiscordBotListApi _api;
+        private readonly AuthDiscordBotListApi _api;
         private readonly Credentials _cred;
 
         public UnitTests()
         {
             _cred = Credentials.LoadFromEnv();
-            _api = new AuthV1DiscordBotListApi(_cred.BotId, _cred.Token);
-        }
-
-        [Fact]
-        public async Task HasVotedTestAsync()
-        {
-            Assert.False(await _api.HasVotedAsync(0));
+            _api = new AuthDiscordBotListApi(_cred.BotId, _cred.Token);
         }
 
         [Fact]
         public async Task TaskIsWeekendTestAsync()
         {
             await _api.IsWeekendAsync();
-        }
-
-        [Fact]
-        public async Task TaskGetVotersTestAsync()
-        {
-            Assert.NotNull(await _api.GetVotersAsync());
-        }
-
-        [Fact]
-        public async Task GetBotTestAsync()
-        {
-            var botId = 1026525568344264724U;
-            var bot = await _api.GetBotAsync(botId);
-            Assert.NotNull(bot);
-            Assert.Equal(botId, bot.Id);
-        }
-
-        [Fact]
-        public async Task GetMeTestAsync()
-        {
-            Assert.NotNull(await _api.GetMeAsync());
-        }
-
-        [Fact]
-        public async Task GetBotsTestAsync()
-        {
-            var bots = await _api.GetBotsAsync();
-
-            Assert.NotNull(bots);
-            Assert.NotEmpty(bots.Items);
         }
 
         [Fact]
