+++ DiscordBotsList.Api/DiscordBotsList.Api.csproj
@@ -16,7 +16,7 @@
         <AssemblyName>DiscordBotsList.Api</AssemblyName>
         <RootNamespace>DiscordBotsList.Api</RootNamespace>
         <Product>DiscordBotsList.Api</Product>
-        <Version>1.6.0</Version>
+        <Version>2.0.0</Version>
         <PackageLicenseFile>LICENSE</PackageLicenseFile>
     </PropertyGroup>
