--- DiscordBotsList.Api/Internal/BotReviews.cs
@@ -1,13 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace DiscordBotsList.Api.Internal
-{
-    public class BotReviews
-    {
-        [JsonPropertyName("averageScore")]
-        public double AverageScore { get; internal set; }
-
-        [JsonPropertyName("count")]
-        public int Count { get; internal set; }
-    }
-}
