--- DiscordBotsList.Api/Internal/HasVotedObject.cs
@@ -1,9 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace DiscordBotsList.Api.Internal
-{
-    internal class HasVotedObject
-    {
-        [JsonPropertyName("voted")] public int? HasVoted { get; set; }
-    }
-}
