--- DiscordBotsList.Api/Internal/SelfBot.cs
@@ -1,39 +1,0 @@
-﻿using DiscordBotsList.Api.Objects;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace DiscordBotsList.Api.Internal
-{
-    internal class SelfBot : Bot, IDblSelfBot
-    {
-        public async Task<List<IDblEntity>> GetVotersAsync(int page = 1)
-        {
-            return await ((AuthDiscordBotListApi)api).GetVotersAsync(page);
-        }
-
-        public async Task<bool> HasVotedAsync(ulong userId)
-        {
-            return await ((AuthDiscordBotListApi)api).HasVotedAsync(userId);
-        }
-
-        public async Task<bool> IsWeekendAsync()
-        {
-            return await ((AuthDiscordBotListApi)api).IsWeekendAsync();
-        }
-
-        public async Task UpdateStatsAsync(int guildCount)
-        {
-            await ((AuthDiscordBotListApi)api).UpdateStatsAsync(guildCount);
-        }
-
-        public async Task UpdateStatsAsync(int[] shards)
-        {
-            await ((AuthDiscordBotListApi)api).UpdateStatsAsync(0, shards.Length, shards);
-        }
-
-        public async Task UpdateStatsAsync(int shardCount, int totalShards, params int[] shards)
-        {
-            await ((AuthDiscordBotListApi)api).UpdateStatsAsync(shardCount, totalShards, shards);
-        }
-    }
-}
